Initial commit
This commit adds the first lines of code for boomplay, the audio player
backend which will be used by boomboxd to play tracks and system sounds.

The original idea was to use mpg123 for this, but the remote control
interface was lacking some functionality. Turns out that with the help
of SDL and libmpg123 it's quite easy to hack a small mp3 player, so I
ventured down that road instead.